PR comments addressed
